backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.126
botocore==1.34.126
celery==5.4.0
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.14
distro==1.9.0
Django==5.0.6
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-jazzmin==3.0.0
django-rest-swagger==2.2.0
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-gis==1.0
drf-yasg==1.21.7
icalendar==5.0.12
idna==3.7
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
MarkupSafe==2.1.5
netaddr==1.3.0
openapi-codec==1.3.2
packaging==24.1
pillow==10.3.0
prompt_toolkit==3.0.46
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
recurring-ical-events==2.2.1
redis==5.0.1
reportlab==4.2.0
requests==2.32.3
s3transfer==0.10.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
wrapt==1.16.0
x-wr-timezone==0.0.7
yookassa==3.1.0
